extend DataExplorerFixture in tests
